triton/requirements.txt
torch
torchvision
python-dotenv